Updated to pull the current storage billing model and factor this into the recommendations.
Also revamped the verbiage around recommendations to take into account
what the current billing model is.

Note that in some cases if a customer switched to physical storage before
around late August 2023 the INFORMATION_SCHEMA SCHEMATA view may not
contain the update as Google did not backfill that data for customers
that changed.

See this bug in Google's bugtracker where I documented this:
https://issuetracker.google.com/issues/296794707